add failing test for job assertStart
I don't understand the code in #541
the quantizing logic adds the division remainder instead of
what i would expect, subtracting it.
I added a test that demonstrates the result I expected,
I'ld like @woodsaj to add a commit that either fixes the test and its comments,
or the logic, whichever is more appropriate,
and then we can merge this branch.