only update the stateCheck field in the DB every second run.
The lastCheck column of the Check table in the DB was being updated
on every check.  The primary reason for this is to allow us to identify
when a check's state is stale and no longer being updated. Hitting the
DB with this many updates is expensive. This change updates the logic
to only update the DB every second check, halfing the number of DB
updates that need to be performed.